Add type convert to save TimeUnits in database, don't schedule notification until item is checked completed

+++ app/src/main/java/com/dalydays/android/reminderlist/data/db/ToDoItem.kt
@@ -12,6 +12,6 @@ data class ToDoItem(
         @ColumnInfo(name = "description") var description: String,
         @ColumnInfo(name = "completed") var completed: Boolean = false,
         @ColumnInfo(name = "scheduled") var scheduled: Boolean = false,
-        @ColumnInfo(name = "duration") var duration: Int? = null,
+        @ColumnInfo(name = "duration") var duration: Long? = null,
         @ColumnInfo(name = "time_unit") var timeUnit: TimeUnit? = null
 )
